requirements.txt
Pillow==10.4.0
protobuf==5.28.2
python-dotenv==1.0.1
selenium==4.25.0
tkinterdnd2==0.4.2
